Extract WordPress Core fuzz validator
